backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
motor==3.3.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
scikit-learn==1.3.2 